moved user auth link to its own folder
