+++ src/core/authentication.agent.info/agent.info.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AgentInfoCacheService } from './agent.info.cache.service';
 import { AgentInfoService } from './agent.info.service';
-import { UserAuthenticationLinkModule } from '@domain/community/user/user.authentication.link.module';
+import { UserAuthenticationLinkModule } from '@domain/community/user-authentication-link/user.authentication.link.module';
 @Module({
   imports: [UserAuthenticationLinkModule],
   providers: [AgentInfoService, AgentInfoCacheService],
