+++ test/unit/domain/community/user/user.authentication.link.service.spec.ts
@@ -1,12 +1,12 @@
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
 import { LogContext } from '@common/enums';
 import { UserAlreadyRegisteredException } from '@common/exceptions';
-import { UserAuthenticationLinkService } from '@domain/community/user/user.authentication.link.service';
+import { UserAuthenticationLinkService } from '@domain/community/user-authentication-link/user.authentication.link.service';
 import {
   UserAuthenticationLinkMatch,
   UserAuthenticationLinkOutcome,
   UserAuthenticationLinkResult,
-} from '@domain/community/user/user.authentication.link.types';
+} from '@domain/community/user-authentication-link/user.authentication.link.types';
 import { LoggerService } from '@nestjs/common';
 
 const createAgentInfo = (overrides: Partial<AgentInfo> = {}) => {
